+++ packages/component-library/src/components.d.ts
@@ -59,6 +59,20 @@ export namespace Components {
          */
         "headerText": string;
     }
+    interface PlmgSvgIcon {
+        /**
+          * Define icon's color.  Can be any valid CSS color value.  By default, the icon will have the same color as the parent's element.
+         */
+        "color": string | undefined;
+        /**
+          * Define icon by its name. Name must be one of the existing icon: https://components.ducky.eco/?path=/story/component-svgicon--all-icons  Default: NULL
+         */
+        "icon": string;
+        /**
+          * Define the icon's size.  Allowed values: <value><unit>  Examples: - 1em - 42px  Default: 1em
+         */
+        "size": string;
+    }
 }
 declare global {
     interface HTMLPlmgButtonElement extends Components.PlmgButton, HTMLStencilElement {
@@ -73,9 +87,16 @@ declare global {
         prototype: HTMLPlmgCardElement;
         new (): HTMLPlmgCardElement;
     };
+    interface HTMLPlmgSvgIconElement extends Components.PlmgSvgIcon, HTMLStencilElement {
+    }
+    var HTMLPlmgSvgIconElement: {
+        prototype: HTMLPlmgSvgIconElement;
+        new (): HTMLPlmgSvgIconElement;
+    };
     interface HTMLElementTagNameMap {
         "plmg-button": HTMLPlmgButtonElement;
         "plmg-card": HTMLPlmgCardElement;
+        "plmg-svg-icon": HTMLPlmgSvgIconElement;
     }
 }
 declare namespace LocalJSX {
@@ -131,9 +152,24 @@ declare namespace LocalJSX {
          */
         "headerText"?: string;
     }
+    interface PlmgSvgIcon {
+        /**
+          * Define icon's color.  Can be any valid CSS color value.  By default, the icon will have the same color as the parent's element.
+         */
+        "color"?: string | undefined;
+        /**
+          * Define icon by its name. Name must be one of the existing icon: https://components.ducky.eco/?path=/story/component-svgicon--all-icons  Default: NULL
+         */
+        "icon"?: string;
+        /**
+          * Define the icon's size.  Allowed values: <value><unit>  Examples: - 1em - 42px  Default: 1em
+         */
+        "size"?: string;
+    }
     interface IntrinsicElements {
         "plmg-button": PlmgButton;
         "plmg-card": PlmgCard;
+        "plmg-svg-icon": PlmgSvgIcon;
     }
 }
 export { LocalJSX as JSX };
@@ -142,6 +178,7 @@ declare module "@stencil/core" {
         interface IntrinsicElements {
             "plmg-button": LocalJSX.PlmgButton & JSXBase.HTMLAttributes<HTMLPlmgButtonElement>;
             "plmg-card": LocalJSX.PlmgCard & JSXBase.HTMLAttributes<HTMLPlmgCardElement>;
+            "plmg-svg-icon": LocalJSX.PlmgSvgIcon & JSXBase.HTMLAttributes<HTMLPlmgSvgIconElement>;
         }
     }
 }